Fixed pyyaml security bug
