+++ src/autopilot/config_reader.py
@@ -7,7 +7,7 @@ def read_gain(path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         gain = conf["rudder gain"]
 
     return eval(gain)
@@ -18,7 +18,7 @@ def read_interval(path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         helm_interval = conf["autohelm interval"]
 
     return eval(helm_interval)
@@ -29,7 +29,7 @@ def lat_tol(path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         helm_interval = conf["longitude tolerance"]
 
     return eval(helm_interval)
@@ -40,7 +40,7 @@ def long_tol(path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         helm_interval = conf["latitude tolerance"]
 
     return eval(helm_interval)
