+++ src/boat/config_reader.py
@@ -2,12 +2,12 @@ import yaml
 import os
 
 
-def read_interval(path=None):
+def upwind_angle(path=None):
     """Reads the read interval from config.yml"""
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
         conf = yaml.full_load(yml)
-        helm_interval = conf["nav interval"]
+        angle = conf["upwind angle"]
 
-    return eval(helm_interval)
+    return angle
