+++ src/rc_input/config_reader.py
@@ -10,7 +10,7 @@ def read_pin_config(mock_bbio=None, path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         if mock_bbio is None:
             pins = {
                 "RUDDER": make_pin(conf["pins"]["RUDDER"]),
@@ -38,7 +38,7 @@ def read_interval(path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         interval = conf["read interval"]
 
     return eval(interval)
