+++ src/rudder/config_reader.py
@@ -10,7 +10,7 @@ def build_pin_from_config(path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         pin = make_pin(conf["pins"]["RUDDER"])
 
     return pin
@@ -21,7 +21,7 @@ def read_servo_config(path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         servo_config = conf["servos"]["MAIN"]
 
     return servo_config
@@ -32,7 +32,7 @@ def read_rudder_config(path=None):
     if path is None:
         path = os.path.dirname(os.path.abspath(__file__))
     with open(path + "/config.yml", "r") as yml:
-        conf = yaml.load(yml)
+        conf = yaml.full_load(yml)
         rudder_config = conf["rudder"]
 
     return rudder_config
